cast the return of unravel_index of numpy to be the same as its input

+++ ivy/functional/backends/numpy/experimental/statistical.py
@@ -50,7 +50,7 @@ def unravel_index(
     *,
     out: Optional[np.ndarray] = None,
 ) -> np.ndarray:
-    return np.unravel_index(indices, shape)
+    return np.unravel_index(indices, shape).astype(indices.dtype)
 
 
 def quantile(
